requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pdfminer.six==20231228
python-docx==1.1.2
pydantic==2.8.2
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0

